manual cleanup of sqlite vector implementation
