--- sqlite/lib/vector_handling/semantic-search/index.js
@@ -1,13 +1,0 @@
-// Main exports for semantic search functionality
-const {
-  embedding,
-} = require('./lib/embeddings.js')
-
-const {
-  createSession
-} = require('./lib/embedding.js')
-
-module.exports = {
-  embedding,
-  createSession
-}
